Renamed team_src and team_headers
