+++ .settings/org.eclipse.core.resources.prefs
@@ -11,6 +11,10 @@ encoding//Debug/Magic\ CAN\ Node\ 2/subdir_rules.mk=UTF-8
 encoding//Debug/Magic\ CAN\ Node\ 2/subdir_vars.mk=UTF-8
 encoding//Debug/Magic\ CAN\ Node\ 2/ti_src/subdir_rules.mk=UTF-8
 encoding//Debug/Magic\ CAN\ Node\ 2/ti_src/subdir_vars.mk=UTF-8
+encoding//Debug/RC\ Headers/subdir_rules.mk=UTF-8
+encoding//Debug/RC\ Headers/subdir_vars.mk=UTF-8
+encoding//Debug/RC\ Source/subdir_rules.mk=UTF-8
+encoding//Debug/RC\ Source/subdir_vars.mk=UTF-8
 encoding//Debug/cmd/subdir_rules.mk=UTF-8
 encoding//Debug/cmd/subdir_vars.mk=UTF-8
 encoding//Debug/lib/subdir_rules.mk=UTF-8
