+++ Magic CAN Node 2/ccsObjs.opt
@@ -1,0 +1,1 @@
+"C:/Users/Chris/Documents/BuckeyeCurrent/MCN_Library/Debug/ti_src/DSP2803x_GlobalVariableDefs.obj" "./main.obj" "./ti_src/DSP2803x_CodeStartBranch.obj" "./ti_src/DSP2803x_DefaultIsr.obj" "./MCN2 Source/Init.obj" "./MCN2 Source/PowerDown.obj" "./MCN2 Source/SensorCov.obj" "./MCN2 Source/can.obj" "./MCN2 Source/clock.obj" "../lib/2803x_FlashAPI_BootROMSymbols.lib" "../lib/Flash2803x_API_V100.lib" "../cmd/DSP2803x_Headers_nonBIOS.cmd" "../cmd/F28035.cmd" -l"libc.a" -l"C:\Users\Chris\Documents\BuckeyeCurrent\MCN_Library\Debug\MCN_Library.lib"
