+++ Magic CAN Node 2/cmd/subdir_rules.mk
@@ -1,0 +1,6 @@
+################################################################################
+# Automatically-generated file. Do not edit!
+################################################################################
+
+# Each subdirectory must supply rules for building sources it contributes
+
