+++ Magic CAN Node 2/cmd/subdir_vars.mk
@@ -1,0 +1,10 @@
+################################################################################
+# Automatically-generated file. Do not edit!
+################################################################################
+
+# Add inputs and outputs from these tool invocations to the build variables 
+CMD_SRCS += \
+../cmd/DSP2803x_Headers_nonBIOS.cmd \
+../cmd/F28035.cmd 
+
+
