+++ Magic CAN Node 2/lib/subdir_vars.mk
@@ -1,0 +1,10 @@
+################################################################################
+# Automatically-generated file. Do not edit!
+################################################################################
+
+# Add inputs and outputs from these tool invocations to the build variables 
+LIB_SRCS += \
+../lib/2803x_FlashAPI_BootROMSymbols.lib \
+../lib/Flash2803x_API_V100.lib 
+
+
