+++ Magic CAN Node 2/objects.mk
@@ -1,0 +1,8 @@
+################################################################################
+# Automatically-generated file. Do not edit!
+################################################################################
+
+USER_OBJS :=
+
+LIBS := -l"libc.a" -l"C:/Users/Chris/Documents/BuckeyeCurrent/MCN_Library/Debug/MCN_Library.lib"
+
