+++ Magic CAN Node 2/sources.mk
@@ -1,0 +1,111 @@
+################################################################################
+# Automatically-generated file. Do not edit!
+################################################################################
+
+O_SRCS := 
+CPP_SRCS := 
+K_SRCS := 
+LD_SRCS := 
+S67_SRCS := 
+LDS_SRCS := 
+CMD_SRCS := 
+CXX_SRCS := 
+CMD_UPPER_SRCS := 
+C43_SRCS := 
+S55_SRCS := 
+LD_UPPER_SRCS := 
+C62_SRCS := 
+S_UPPER_SRCS := 
+A_SRCS := 
+SA_SRCS := 
+C55_SRCS := 
+LDS_UPPER_SRCS := 
+C_UPPER_SRCS := 
+INO_SRCS := 
+OBJ_SRCS := 
+S62_SRCS := 
+LIB_SRCS := 
+PDE_SRCS := 
+ASM_SRCS := 
+ASM_UPPER_SRCS := 
+C++_SRCS := 
+CLA_SRCS := 
+S??_SRCS := 
+C_SRCS := 
+C67_SRCS := 
+S_SRCS := 
+S43_SRCS := 
+OPT_SRCS := 
+C64_SRCS := 
+CC_SRCS := 
+C??_SRCS := 
+S64_SRCS := 
+OBJS := 
+S_DEPS := 
+S_UPPER_DEPS := 
+S62_DEPS := 
+C64_DEPS := 
+ASM_DEPS := 
+CC_DEPS := 
+S55_DEPS := 
+C67_DEPS := 
+CLA_DEPS := 
+C??_DEPS := 
+CPP_DEPS := 
+S??_DEPS := 
+C_DEPS := 
+C62_DEPS := 
+EXE_OUTPUTS := 
+CXX_DEPS := 
+C++_DEPS := 
+ASM_UPPER_DEPS := 
+K_DEPS := 
+C43_DEPS := 
+INO_DEPS := 
+S67_DEPS := 
+SA_DEPS := 
+S43_DEPS := 
+OPT_DEPS := 
+PDE_DEPS := 
+S64_DEPS := 
+C_UPPER_DEPS := 
+C55_DEPS := 
+CPP_DEPS__QUOTED := 
+C67_DEPS__QUOTED := 
+INO_DEPS__QUOTED := 
+C??_DEPS__QUOTED := 
+S_UPPER_DEPS__QUOTED := 
+CLA_DEPS__QUOTED := 
+ASM_UPPER_DEPS__QUOTED := 
+C62_DEPS__QUOTED := 
+CXX_DEPS__QUOTED := 
+EXE_OUTPUTS__QUOTED := 
+S67_DEPS__QUOTED := 
+C_DEPS__QUOTED := 
+C_UPPER_DEPS__QUOTED := 
+OPT_DEPS__QUOTED := 
+S_DEPS__QUOTED := 
+K_DEPS__QUOTED := 
+S??_DEPS__QUOTED := 
+C64_DEPS__QUOTED := 
+C++_DEPS__QUOTED := 
+OBJS__QUOTED := 
+CC_DEPS__QUOTED := 
+S43_DEPS__QUOTED := 
+S55_DEPS__QUOTED := 
+SA_DEPS__QUOTED := 
+C55_DEPS__QUOTED := 
+PDE_DEPS__QUOTED := 
+C43_DEPS__QUOTED := 
+S62_DEPS__QUOTED := 
+ASM_DEPS__QUOTED := 
+S64_DEPS__QUOTED := 
+
+# Every subdirectory with source files must be described here
+SUBDIRS := \
+. \
+ti_src \
+lib \
+cmd \
+MCN2\ Source \
+
