+++ Magic CAN Node 2/subdir_vars.mk
@@ -1,0 +1,27 @@
+################################################################################
+# Automatically-generated file. Do not edit!
+################################################################################
+
+# Add inputs and outputs from these tool invocations to the build variables 
+OBJ_SRCS += \
+C:/Users/Chris/Documents/BuckeyeCurrent/MCN_Library/Debug/ti_src/DSP2803x_GlobalVariableDefs.obj 
+
+C_SRCS += \
+../main.c 
+
+OBJS += \
+./main.obj 
+
+C_DEPS += \
+./main.pp 
+
+C_DEPS__QUOTED += \
+"main.pp" 
+
+OBJS__QUOTED += \
+"main.obj" 
+
+C_SRCS__QUOTED += \
+"../main.c" 
+
+
