+++ RC Headers/all.h
@@ -11,35 +11,35 @@
 #include "template_all.h"
 #include "DSP28x_Project.h"     // Device Headerfile and Examples Include File
 #include "stopwatch_template.h"
-#include "op.h"
 #include "sys_op.h"
-#include "data.h"
 #include "sys_data.h"
-#include "Init.h"
 #include "Init_template.h"
 #include "can_template.h"
-#include "SensorCov.h"
 #include "SensorCov_template.h"
 #include "boot_template.h"
-#include "PowerDown.h"
 #include "PowerDown_template.h"
 #include "common_template.h"
-#include "main.h"
-#include "clock.h"
 #include "clock_template.h"
-#include "can.h"
 #include <stdlib.h>
 #include <math.h>
 #include "adc_template.h"
 #include "gpio_template.h"
 #include "DSP2803x_GlobalPrototypes.h"
 #include "Flash2803x_API_Library.h"
-#include "IQmathLib.h"
-#include "EMA.h"
-#include "Stack.h"
 #include <stdbool.h>
-#include "SafetyVar.h"
-#include "GPIO.h"
+#include "../RC Headers/can.h"
+#include "../RC Headers/clock.h"
+#include "../RC Headers/data.h"
+#include "../RC Headers/EMA.h"
+#include "../RC Headers/GPIO.h"
+#include "../RC Headers/Init.h"
+#include "../RC Headers/IQmathLib.h"
+#include "../RC Headers/main.h"
+#include "../RC Headers/op.h"
+#include "../RC Headers/PowerDown.h"
+#include "../RC Headers/SafetyVar.h"
+#include "../RC Headers/SensorCov.h"
+#include "../RC Headers/Stack.h"
 
 
 extern sys_ops_struct sys_ops;
