+++ RC Source/SafetyVar.c
@@ -5,7 +5,7 @@
  *      Author: David
  */
 
-#include "all.h"
+#include "../RC Headers/all.h"
 
 void SafetyVar_NewValue(SafetyVar32_t *ptr, Uint32 value){
