+++ RC Source/Stack.c
@@ -6,7 +6,7 @@
  */
 
 
-#include "all.h"
+#include "../RC Headers/all.h"
 
 extern void _stack;
 extern void _STACK_SIZE;
